Tweak probabilities of adding a shape and mutating existing shapes
The idea here is that the probability of adding a shape decreases
as the size of the genome decreases, until the probability hits
zero when we have hit a maximum size.

When it comes to mutating an existing shape, the idea is to
reduce the probability as the genome gets bigger such that the
expected number of shapes mutated is constant.