Pass evolverSettings into every method that needs it
The idea is to make settings stop being a singleton and instead make it a thing
which is owned by the evolver that uses it. I want it to be possible to read
the settings from a DB table or from the DudeStore server so you can have
different evolvers going with different settings.